% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeREMBO.R
\name{activeREMBO}
\alias{activeREMBO}
\title{Random Embedding Bayesian Optimization}
\usage{
activeREMBO(
  par,
  fn,
  lower,
  upper,
  budget,
  ...,
  highDimGP,
  useAScrit = FALSE,
  ASalternative = FALSE,
  homcontrol = list(beta0 = 0, covtype = "Matern5_2", g = sqrt(.Machine$double.eps)),
  kmcontrol = list(covtype = "matern5_2", iso = TRUE, covreestim = TRUE, formula = ~1),
  control = list(Atype = "isotropic", reverse = TRUE, bxsize = NULL, testU = TRUE,
    standard = FALSE, maxitOptA = 100, lightreturn = FALSE, warping = "Psi", designtype =
    "unif", tcheckP = 1e-04, roll = F, returnAS = TRUE, plotting = FALSE, inneroptim =
    "pso", popsize = 80, gen = 40),
  init = NULL
)
}
\arguments{
\item{par}{vector whose length is used to define the initial DoE or just the low dimension}

\item{fn}{function to be minimized over}

\item{lower, upper}{bounds for optimization}

\item{budget}{total number of calls to the objective function}

\item{...}{additional parameters of fn}

\item{highDimGP}{should the GP use true design values?}

\item{useAScrit}{should an optimization on a sequential active subspace identification criterion be performed in the orthogonal?}

\item{ASalternative}{alternate between AS criterion and EI}

\item{kmcontrol}{an optional list of control parameters to be passed to the \code{\link[DiceKriging]{km}} model:
\code{iso}, \code{covtype}, \code{formula}. In addition, boolean \code{codereestim} is passed to \code{\link[DiceKriging]{update.km}}}

\item{control}{an optional list of control parameters. See "Details"}

\item{init}{optional list with elements \code{Amat} to provide a random matrix, \code{low_dim_design} for an initial design in the low-dimensional space and \code{fvalues} for the corresponding response.
When passing initial response values, care should be taken that the mapping with \code{Amat} of the design actually correspond to high-dimensional designs giving \code{fvalues}.}
}
\description{
REMBO for unconstrained problems, with changing embedding
}
\details{
Options available from \code{control} are:
\itemize{
\item \code{Atype} see \code{\link[RRembo]{selectA}};
\item \code{reverse} if \code{TRUE}, use the new mapping from the zonotope,
 otherwise the original mapping with convex projection;
\item \code{bxsize} scalar controling the box size in the low-dimensional space;
\item \code{testU} with the regular mapping, set to \code{TRUE} to check that points are in U (to avoid non-injectivity);
\item \code{standard} for using settings of the original REMBO method;
\item \code{maxitOptA} if \code{Atype} is \code{optimized}, number of optimization iterations;
\item \code{lightreturn} only returns \code{par} and \code{value};
\item \code{warping} either \code{"standard"} for kY, \code{"kX"} or \code{"Psi"};
\item \code{designtype} one of "\code{LHS}", "\code{maximin}" and "\code{unif}",
 see \code{\link[RRembo]{designZ}} or \code{\link[RRembo]{designU}};
\item \code{tcheckP} minimal distance to an existing solution, see \code{\link[GPareto]{checkPredict}}
\item \code{roll} to alternate between optimization methods;
\item \code{inneroptim} optimization method for EI
\item \code{popsize, gen} population size and number of optimization generations of EI
}
}
\examples{
\dontrun{
set.seed(42)
library(rgl)
library(DiceKriging)

lowd <- 2
highD <- 25

ntest <- 50
maxEval <- 100

res <- rep(0, ntest)
for(i in 1:ntest){
  #ii <- sample(1:highD, 2)
  ii <- c(1,2)
  branin_mod2 <- function(X){
    if(is.null(nrow(X))) X <- matrix(X, nrow = 1)
    X <- X[, c(ii[1], ii[2]), drop = FALSE]
    return(apply(X, 1, branin))
  }
  sol <- activeREMBO(par = rep(NA, lowd), branin_mod2, lower = rep(0, highD),
                   upper = rep(1, highD), budget = maxEval, highDimGP = TRUE)
  res[i] <- sol$value
  cat(sol$value, " ", i, "\n")
}

plot(res - 0.397887, type = "b")
boxplot(res - 0.397887)
}
}
\references{
O. Roustant, D. Ginsbourger & Y. Deville, DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization Journal of Statistical Software, 2012, 51, 1-55\cr \cr
Z. Wang, F. Hutter, M. Zoghi, D. Matheson, N. de Freitas (2016), Bayesian Optimization in a Billion Dimensions via Random Embeddings, JAIR. \cr \cr
M. Binois, D. Ginsbourger, O. Roustant (2015), A Warped Kernel Improving Robustness in Bayesian Optimization Via Random Embeddings, Learning and Intelligent Optimization, Springer \cr \cr
M. Binois, D. Ginsbourger, O. Roustant (2018), On the choice of the low-dimensional domain for global optimization via random embeddings, arXiv:1704.05318 \cr \cr
M. Binois (2015), Uncertainty quantification on Pareto fronts and high-dimensional strategies in Bayesian optimization, with applications in multi-objective automotive design, PhD thesis, Mines Saint-Etienne.
}
\author{
Mickael Binois
}
