% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{designU}
\alias{designU}
\title{Select design in U}
\usage{
designU(p, A, bxsize, type = "maximin", standard = FALSE)
}
\arguments{
\item{p}{number of points of the design}

\item{A}{random embedding matrix}

\item{bxsize}{scaler bounds on the domain, i.e., bxsize x [-1,1]^D}

\item{type}{design type, one of "LHS", "maximin" and 'unif'}

\item{standard}{standard REMBO approach}
}
\description{
Select design in U
}
\examples{
## Example of designs in U
set.seed(42)
d <- 2; D <- 5

A <- selectA(d, D, type = 'optimized')
size <- 5 # box size of Y
ntest <- 10000
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)

inU <- testU(Y, A)
colors <- rep('black', ntest)
colors[inU] <- 'green'
plot(Y, col = colors, pch = 20, cex = 0.5)

p <- 20
designs1 <- designU(p, A, size)
designs2 <- designU(p, A, size, type = 'LHS')

points(designs1, col = 'red', pch = 20)
points(designs2, col = 'blue', pch = 20)
legend("topright", legend = c("LHS", "Maximin LHS"), col = c("blue", "red"), pch = c(20,20))

}
\references{
M. Binois, D. Ginsbourger, O. Roustant (2018), On the choice of the low-dimensional domain for global optimization via random embeddings, arXiv:1704.05318 \cr \cr
M. Binois (2015), Uncertainty quantification on Pareto fronts and high-dimensional strategies in Bayesian optimization, with applications in multi-objective automotive design, PhD thesis, Mines Saint-Etienne.
}
\author{
Mickael Binois
}
