% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{designZ}
\alias{designZ}
\title{Select design in Z}
\usage{
designZ(p, pA, bxsize, type = "maximin")
}
\arguments{
\item{p}{number of points of the design}

\item{pA}{orthogonal projection onto Ran(A) matrix}

\item{bxsize}{bounds of the domain}

\item{type}{design type, one of "LHS", "maximin", "unif"}
}
\description{
Select design in Z
}
\examples{
## Example of designs in Z
set.seed(42)
d <- 2; D <- 5

A <- selectA(d, D, type = 'optimized')
size <- sqrt(D) # box size of Z
ntest <- 10000
Z <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)

inZ <- testZ(Z, t(A))
colors <- rep('black', ntest)
colors[inZ] <- 'green'
plot(Z, col = colors, pch = 20, cex = 0.5)

p <- 20
designs1 <- designZ(p, t(A), size)
designs2 <- designZ(p, t(A), size, type = 'LHS')

points(designs1, col = 'red', pch = 20)
points(designs2, col = 'blue', pch = 20)
legend("topright", legend = c("LHS", "Maximin LHS"), col = c("blue", "red"), pch = c(20,20))

}
\references{
M. Binois, D. Ginsbourger, O. Roustant (2018), On the choice of the low-dimensional domain for global optimization via random embeddings, arXiv:1704.05318 \cr \cr
M. Binois (2015), Uncertainty quantification on Pareto fronts and high-dimensional strategies in Bayesian optimization, with applications in multi-objective automotive design, PhD thesis, Mines Saint-Etienne.
}
\author{
Mickael Binois
}
