% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mapZX}
\alias{mapZX}
\title{Alternative mapping}
\usage{
mapZX(z, A, eps = 1e-06, Amat = NULL, Aind = NULL)
}
\arguments{
\item{z}{matrix of low-dimensional coordinates in the zonotope, one point per row}

\item{A}{random embedding matrix, such that \code{t(A)A = Id}}

\item{eps}{to avoid some numerical issues with quadratic programming}

\item{Aind, Amat}{optional matrices to be passed to \code{\link[quadprog]{solve.QP.compact}}}
}
\description{
Map from zonotope Z to the convex projection of A Y (E)
}
\details{
Numerical problems may occur, then the equality constraints are slightly relaxed.
}
\examples{
## Example of forward - backward mapping
d <- 5; D <- 26
A <- selectA(d, D, type = 'Gaussian')

n <- 10000
size <- 10
Y <- size * (2 * matrix(runif(n * d), n) - 1)
X <- randEmb(Y, A)
Z <- ortProj(X, t(A))
# Errors are catched and the problem slightly relaxed
Xback <- mapZX(Z, A)

print(max(abs(Xback - X)))
print(mean(Xback - X))
}
\references{
M. Binois, D. Ginsbourger, O. Roustant (2018), On the choice of the low-dimensional domain for global optimization via random embeddings, arXiv:1704.05318 \cr
}
\author{
Mickael Binois
}
