% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{ortProj}
\alias{ortProj}
\title{Orthogonal projection on Ran(A)}
\usage{
ortProj(x, pA)
}
\arguments{
\item{x}{matrix of high-dimensional coordinates, one point per row}

\item{pA}{matrix giving the coordinates of the orthogonal projection onto Ran(A)}
}
\value{
z matrix of coordinates of the orthogonal projection onto Ran(A)
}
\description{
Orthogonal projection on the subspace spanned by A
}
\details{
It is assumed that rows of pA are orthonormal, such that this linear transformation gives the coordinates on Ran(A) with an orthonormal basis
}
\examples{
## Example of orthogonal projection
d <- 2; D <- 6
A1 <- selectA(d, D, type = 'Gaussian')
A2 <- selectA(d, D, type = 'isotropic')
A3 <- selectA(d, D, type = 'optimized')

n <- 10000
size <- 10
Y <- size * (2 * matrix(runif(n * d), n) - 1)

Z1 <- ortProj(randEmb(Y, A1), t(A1))
Z2 <- ortProj(randEmb(Y, A2), t(A2))
Z3 <- ortProj(randEmb(Y, A3), t(A3))

par(mfrow = c(1, 3))
plot(Z1, asp = 1)
plot(Z2, asp = 1)
plot(Z3, asp = 1)

par(mfrow = c(1, 1))
}
\author{
Mickael Binois
}
