% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpings.R
\name{Psi_Y}
\alias{Psi_Y}
\title{Warping psi for points in Y}
\usage{
Psi_Y(y, A)
}
\arguments{
\item{y}{matrix of low dimensional coordinates, one point per row}

\item{A}{random embedding matrix}
}
\description{
Warping psi (original low dimensional space)
}
\examples{
set.seed(42)
d <- 2; D <- 5

A <- selectA(d, D, type = 'optimized')
size <- 5 # box size of Y
ntest <- 10000
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)
Z <- Psi_Y(Y, A)
plot(Z)
}
\references{
M. Binois, D. Ginsbourger, O. Roustant (2015), A Warped Kernel Improving Robustness in Bayesian Optimization Via Random Embeddings, Learning and Intelligent Optimization, Springer \cr \cr
}
