% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpings.R
\name{Psi_Y_nonort}
\alias{Psi_Y_nonort}
\title{Warping psi for points in Y, non normalized A}
\usage{
Psi_Y_nonort(y, A)
}
\arguments{
\item{y}{matrix of low dimensional coordinates, one point per row}

\item{A}{random embedding matrix with non-othogonal columns}
}
\description{
Warping psi for points in Y, non normalized A
}
\examples{
set.seed(42)
d <- 2; D <- 5
library(MASS)
A <- selectA(d, D, type = 'standard')
size <- 5 # box size of Y
ntest <- 10000
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)
Z <- Psi_Y(Y, A)
plot(Z)
Z2 <- Psi_Y_nonort(Y, A)
points(Z2, col = 'red')
library(far)
B <- orthonormalization(A, basis = FALSE)
Z3 <- Psi_Y(Y, B)
points(Z3, pch = 20)
}
\seealso{
\code{\link[RemboIV]{Psi_Y}} if \code{A} has orthogonal columns
}
