% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpings.R
\name{Psi_Z}
\alias{Psi_Z}
\title{Warping psi for points in the zonotope Z}
\usage{
Psi_Z(z, A, eps = 1e-06)
}
\arguments{
\item{z}{matrix of low dimensional coordinates, one point per row. These points must belong to Z, that can be checked by testZ}

\item{A}{random embedding matrix}

\item{eps}{to avoid some numerical issues with quadratic programming}
}
\description{
Warping psi for the alternative mapping procedure
}
\examples{
# Comparison with between results starting from Y and from Z
set.seed(42)
d <- 2; D <- 5

A <- selectA(d, D, type = 'optimized')
sizeY <- 5 # box size of Y
sizeZ <- sqrt(D) # box size of Z
ntest <- 10000
Y <- sizeY * (2 * matrix(runif(ntest * d), ntest, d) - 1)
X <- randEmb(Y, A)
Z <- ortProj(X, t(A))

PsisY <- Psi_Y(Y, A)
ind <- testZ(Z, t(A), eps = 1e-8) ## Should be all TRUE, up to numerical error
PsisZ <- Psi_Z(Z[ind,], A)
plot(PsisY)
points(PsisZ, col = 'red', pch = 20)
print(max(abs(PsisY - PsisZ)))
}
\references{
M. Binois, D. Ginsbourger, O. Roustant (2018), On the choice of the low-dimensional domain for global optimization via random embeddings, arXiv:1704.05318 \cr
}
