% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{randEmb}
\alias{randEmb}
\title{Random embedding of a low dimensional subspace on the centered hypercube, based on a matrix and convex projection
Random embedding mapping}
\usage{
randEmb(y, A)
}
\arguments{
\item{y}{matrix of low dimensional coordinates, one point per row}

\item{A}{random embedding matrix}
}
\value{
Matrix corresponding to convex projection onto [-1, 1]^D of A * y
}
\description{
Random embedding of a low dimensional subspace on the centered hypercube, based on a matrix and convex projection
Random embedding mapping
}
\examples{
## Small dimensional examples for illustration
d <- 1; D <- 2
set.seed(42)

ntest <- 1000
size <- 1.5 # box size of Y
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)
A <- matrix(rnorm(D *d), D, d)
X <- randEmb(Y, A)
plot(t(A \%*\% t(Y)), type = 'l', ylim = c(-1,1), xlim = c(-1, 1))
points(X, col = 'red', pch = 20)

library(rgl)
d <- 2; D <- 3
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)
A <- matrix(rnorm(D *d), D, d)
X <- randEmb(Y, A)
plot3d(t(A \%*\% t(Y)))
points3d(X, col = 'red', pch = ".")
}
\references{
Z. Wang, F. Hutter, M. Zoghi, D. Matheson, N. de Freitas (2016), Bayesian Optimization in a Billion Dimensions via Random Embeddings, JAIR. \cr \cr
}
