% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{testU}
\alias{testU}
\title{Test if a given point is in U}
\usage{
testU(y, A)
}
\arguments{
\item{y}{matrix corresponding to points to be tested}

\item{A}{random embedding matrix}
}
\description{
Test if a given point is in U
}
\examples{
## Identification of the U set
set.seed(42)
d <- 2; D <- 5

A <- selectA(d, D)
size <- 15 # box size of Y
ntest <- 10000
Y <- size * (2 * matrix(runif(ntest * d), ntest, d) - 1)

inU <- testU(Y, A)
colors <- rep('black', ntest)
colors[inU] <- 'green'
plot(Y, col = colors, pch = 20, cex = 0.5)

}
